refactor: ⚡️ just moving functions around
